requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via emergencybd-backend-private
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
cryptography==45.0.7
    # via emergencybd-backend-private
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.116.1
    # via emergencybd-backend-private
fastapi-cli==0.0.11
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
gunicorn==23.0.0
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via gunicorn
pillow==11.3.0
    # via emergencybd-backend-private
pycparser==2.23 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.11.8
    # via
    #   emergencybd-backend-private
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via emergencybd-backend-private
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via emergencybd-backend-private
pyotp==2.9.0
    # via emergencybd-backend-private
python-dotenv==1.1.1
    # via
    #   emergencybd-backend-private
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   emergencybd-backend-private
    #   fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.37.1
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via sqlmodel
sqlcipher3-binary==0.5.4
    # via emergencybd-backend-private
sqlmodel==0.0.24
    # via emergencybd-backend-private
starlette==0.47.3
    # via fastapi
typer==0.17.4
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via sentry-sdk
uvicorn==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
